requirements.txt
discord.py[voice]
PyNaCl
Flask